ops-agent-mcp/requirements.txt
# Core dependencies
fastmcp==2.12.4

# HTTP client
requests==2.32.3

# Configuration management
pyyaml==6.0.2
python-dotenv>=1.1.0

# Logging
colorlog==6.9.0
